requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bidict==0.23.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
Flask==3.0.2
Flask-SocketIO==5.3.6
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
